Extract engine storyboard operation
storyboard is a bounded media-output operation with frame extraction and optional grid generation. This moves it into engine_storyboard.py while re-exporting storyboard from mcp_video.engine for compatibility.

Constraint: Existing callers import storyboard from mcp_video.engine and expect StoryboardResult behavior unchanged.

Rejected: Move subtitles/watermark with storyboard | those are separate rendering contracts and should be reviewed independently.

Confidence: high

Scope-risk: narrow

Directive: Keep media-output operation extraction small and facade-compatible.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_storyboard.py

Tested: /opt/homebrew/bin/python3 storyboard re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py -k 'storyboard' tests/test_server.py -k 'storyboard' tests/test_cli.py -k 'storyboard' tests/test_e2e.py -k 'storyboard' -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
